refactor(core): extract createMediaLibraryClient into dependency-free util
Importing it from core/store/accessPolicy/fetch pulled that module's
top-level DataLoader/QuickLRU caches into broad UI/upload flows via
client-adapters/assets.ts. Moving the helper to core/util avoids eagerly
initializing policy-fetching infrastructure when only video asset
observation is needed.
